Rename tx-log service to transaction log
