+++ packages/operator/src/services/state-manager/transaction-log.ts
